Refactor session tests into multiple modules
