+++ crates/hotfix/tests/session_test_cases/mod.rs
@@ -1,0 +1,1 @@
+mod heartbeat_tests;
